# generated/swagger-ui/swagger.json.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Project to demonstrate swagger 2.0 integration with spring boot",
    "version" : "v1",
    "title" : "Spring Boot Swagger Integration",
    "termsOfService" : "http://www.example.com/termsOfService",
    "contact" : {
      "name" : "Name",
      "url" : "http://www.example.com",
      "email" : "email@email.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:8081",
  "basePath" : "/swagger",
  "schemes" : [ "http" ],
  "paths" : {
    "/messages/" : {
      "get" : {
        "operationId" : "getMessages",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Message"
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "createMessage",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Message"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Message"
            }
          }
        }
      }
    },
    "/messages/{messageId}" : {
      "get" : {
        "operationId" : "getMessage",
        "parameters" : [ {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Message"
            }
          }
        }
      },
      "put" : {
        "operationId" : "updateMessage",
        "parameters" : [ {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Message"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Message"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "deleteMessage",
        "parameters" : [ {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "securityDefinitions" : {
    "ApiKeyAuth" : {
      "type" : "apiKey",
      "name" : "x-api-key",
      "in" : "header"
    }
  },
  "definitions" : {
    "Message" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string"
        }
      }
    }
  }
}