# src/main/java/com/sanchi/java/swagger/models/Message.java
package com.sanchi.java.swagger.models;

import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;

@ApiModel
public class Message {

  private int id;
  @ApiModelProperty(value = "code", example = "A1")
  private String code;
  @ApiModelProperty(value = "message", example = "My message", required = true)
  private String message;

  public int getId() {
    return id;
  }

  public void setId(int id) {
    this.id = id;
  }

  public String getCode() {
    return code;
  }

  public void setCode(String code) {
    this.code = code;
  }

  public String getMessage() {
    return message;
  }

  public void setMessage(String message) {
    this.message = message;
  }
}